DAO-212 Move search-form files into folder
