+++ src/pages/components/search-form/index.ts
@@ -1,0 +1,1 @@
+export { default } from './search-form';
